chtml/html/testdata/webkit/namespace-sensitivity.dat

#data
<body><table><tr><td><svg><td><foreignObject><span></td>Foo
#errors
#document
| <body>
|   "Foo"
|   <table>
|     <tbody>
|       <tr>
|         <td>
|           <svg svg>
|             <svg td>
|               <svg foreignObject>
|                 <span>
